Rename to grid.py
