+++ src/jaxgym/grid.py
